% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vndirect.R
\name{companyInfo}
\alias{companyInfo}
\title{Company Information}
\usage{
companyInfo(.ticker = NULL)
}
\arguments{
\item{.ticker}{Ticker with 3 characters. Pass this parameter if you want to return information for a specific company.}
}
\value{
A data frame containing company information.
}
\description{
Retrieves company information from the VNDIRECT API for a given ticker.
}
\examples{
# Retrieve information for a specific company
cpn_info <- companyInfo()

}
