% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vndirect.R
\name{getData}
\alias{getData}
\title{REQUEST THE VNDIRECT API}
\usage{
getData(.tickers = NULL, .size = 100)
}
\arguments{
\item{.tickers}{A character vector of stock symbols.}

\item{.size}{Length of historical data.}
}
\value{
A data frame containing historical stock data for all tickers.
}
\description{
Retrieves historical stock data from the VNDIRECT API for multiple tickers.
}
\examples{
getData(c('TPB', 'VCB', 'HCM'), 1000)

}
