% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vndirect.R
\name{vnd_company_info}
\alias{vnd_company_info}
\title{Company information}
\usage{
vnd_company_info(.ticker = character())
}
\arguments{
\item{.ticker}{ticker with 3 characters. Pass this parameter if you want to return all company}
}
\description{
Company information
}
\examples{
tpb_info <- vnd_company_info("TPB")

# Get all company after that filter 

library(dplyr)
company_info <- vnd_company_info() \%>\% 
  filter(floor == 'HOSE') \%>\% 
  filter(indexCode == 'VN30')
}
