% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vndirect.R
\name{vnd_get_data}
\alias{vnd_get_data}
\title{Request the VNDIRECT API}
\usage{
vnd_get_data(.ticker = NULL, .size = 1000)
}
\arguments{
\item{.ticker}{stock symbol}

\item{.size}{length of historical data}
}
\value{
data.frame
}
\description{
Request the VNDIRECT API
}
