% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vndirect.R
\name{vnd_get_list_data}
\alias{vnd_get_list_data}
\title{GET DATA FROM MULTIPLE STOCKS}
\usage{
vnd_get_list_data(.tickers = NULL, .size = 100)
}
\arguments{
\item{.tickers}{list of stock symbols}

\item{.size}{length of historical data}
}
\value{
data.frame
}
\description{
GET DATA FROM MULTIPLE STOCKS
}
\examples{
vnd_get_list_data(c('TPB', 'VCB'))
}
