% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vndirect.R
\name{vnd_get_data}
\alias{vnd_get_data}
\title{REQUEST THE VNDIRECT API}
\usage{
vnd_get_data(.ticker = NULL, .size = 100)
}
\arguments{
\item{.ticker}{stock symbol}

\item{.size}{length of historical data}
}
\value{
data.frame
}
\description{
REQUEST THE VNDIRECT API
}
\examples{
vnd_get_data('TPB', 1000)
}
